lab2/modules/4-module/CreateThumbnail/requirements.txt
# Install into the deployment package on Amazon Linux (see build-lambda-package.sh).
# Lambda already provides boto3; only Pillow must be bundled.
pillow>=11,<13
